vacuumgaugereadout/requirements.txt
argparse
datetime
requests
pyserial
typing